feat(engine): C++ 라우팅 엔진 고도화 Phase A~D (계약·관경물리·다중배관품질·검증)
개발계획서 docs/routing3d_engine_improvement_plan.md 전 단계 완료.
모든 신기능 기본값=기존 동작 → 골든 바이트 불변(ctest 11→13/13).

Phase A (계약·안전성, 골든 불변):
- A1 실패사유 RouteFail enum: astar→SceneResult.fail→C ABI fail_reason→C# RouteResult.Fail
- A2 collect_visited 기본 OFF (뷰어 방문맵/단계탐색 시 opt-in)
- A3 r3d_route_ripup·r3d_copy_blocked >5M셀 ImplicitOccupancy 전환 (Dense 폭발 방지)
- A4 clearance_map narrowing 제거 + 인덱스 계약 문서화 / A5 UTF-8 점검

Phase B (관경 물리):
- B1 per-task 반경 자동 산출 (radius=clamp(ceil(d/cell)-1,0,8)) — 실측 18/20→20/20·18×
- B2 관경 기반 clearance 페널티 (params_for) / B3 FCL 통합=보류(dep-free DLL 보존)

Phase C (다중배관 품질):
- C1 negotiated-congestion(CBS-lite): 평면 rip-up을 연쇄(재귀) rip-up으로 확장 —
  blocker의 blocker까지 bounded depth 재귀 양보(무손실 단조+1·결정적). r3d_set_cbs_depth
- C2 코너 최소반경: 짧은 단관을 경로(셀) 단계에서 충돌검사 하에 흡수(enforce_min_straight).
  비교란 최종 패스(라우팅 점유 불변) — 실측 135→133꺾임. r3d_set_min_straight

Phase D (검증):
- D1 실데이터 회귀 골든 test_realdata (proj1 Exhaust c100 픽스처, 20/20·67600mm·재현성)
- D2 카테고리 골든 test_category (실패사유·대형격자·per-task·CBS·min-straight 무손실/결정적)

뷰어: Engine/Native 래퍼·BuildEngineForRows opt-in 연결·체크박스(협상 라우팅(CBS)·코너 최소반경).
env: R3D_CBS·R3D_MIN_STRAIGHT·R3D_PER_TASK_RADIUS.

+++ cpp/include/routing3d/cost.hpp
@@ -76,14 +76,14 @@ std::vector<int> clearance_map(const Occ& occ, int max_radius, int connectivity 
     std::vector<int> dist(static_cast<size_t>(N), max_radius);
     if (max_radius == 0) {
         for (long long idx = 0; idx < N; ++idx)
-            if (occ.is_blocked(occ.unlin(static_cast<int>(idx)))) dist[static_cast<size_t>(idx)] = 0;
+            if (occ.is_blocked(occ.unlin(idx))) dist[static_cast<size_t>(idx)] = 0;   // A4: 64비트 idx 직접(narrowing 제거).
         return dist;
     }
 
     // 다중소스 BFS: 장애물(dist=0)에서 출발해 연결성 이웃으로 +1 씩 확장(max_radius 상한).
     std::deque<Cell> q;
     for (long long idx = 0; idx < N; ++idx) {
-        Cell c = occ.unlin(static_cast<int>(idx));
+        Cell c = occ.unlin(idx);   // A4: long long idx 직접 전달(int 캐스팅 narrowing 제거).
         if (occ.is_blocked(c)) {
             dist[static_cast<size_t>(idx)] = 0;
             q.push_back(c);
